webhooks plugin, initial commit
This plugin sends HTTP POST request to configured URL(s) when a project
event happens.  The target URL(s) are configurable on project level.
The configuration is inheritable.

This can be used as an alternative to the SSH stream-events.

Motivated by Github's webhooks [1].

Most of code was copied from the high-availability plugin [2].

[1] https://developer.github.com/webhooks/
[2] https://gerrit-review.googlesource.com/#/admin/projects/plugins/high-availability

Change-Id: I902c8d25a25dd54a0b929c6e383837135737e6cb