WIP: Reschedule unsuccessful requests instead of sleeping
Prior to this change the retry logic was implemented at the http client
level and it would simply sleep and retry in a loop. This was suboptimal
as it keeps a thread allocated all the time.

When a web hook can not be executed, reschedule the task to be executed
again after the configured retry interval. This is the same approach
as we have in the replication plugin.

TODO:
* provide tests for the retry logic

Change-Id: I53486cb8752990991763acf9dd83d075b2aa7273