Introduce disableSslVerification configuration parameter
New disableSslVerification parameter was introduced to handle HTTPS SSL
certificate verification. It might be set either globally (in
gerrit.config) with default value 'true' (current behavior is
preserved) or locally per remote (when not specified it defaults to
global).

Implementation details:
1. once HttpClient is built one cannot change its
ConnectionSocketFactory to either verify or ignore SSL certificate
2. introduction 'https+verify' schema was considered but it looked
somewhat hackish as it requires remote URL rewrite prior sending (not
mentioning the fact that down the HttpClient implementation either
'https' or 'http' schema types are only assumed and I was not sure if
could properly trace all places to be changed)
3. depending on effective remote configuration either verifying or
ignoring HttpClient is being used to perform post operation

Change-Id: I52447e670cc0c8c6b113a7234c35d1838ce7c45d
Signed-off-by: Jacek Centkowski <jcentkowski@collab.net>