Documentation: Remove bazel issue
Manual testing showed that the issue mention in documentation is no
longer present in the current version of bazel. Plugin can be build in
tree even after standalone build. The bazel issue was more about
running plugin in-tree testing using filter option after building it
standalone. Second option of plugin in-tree testing (explicit) doesn't
have this issue. So, bazel issue section can be removed as it was
addressing plugin build.

Change-Id: I4a354a91234f2ab6e9292cd770a693ead5867ed9