Avoid triggering webhooks for remote events
The webhooks pluin was unaware of the instance-id of
the events received and caused the trigger of remote
hooks multiple times for the same event, when used in
a multi-site configuration.

Inject the instance-id and avoid running webhooks for
events that are associated with a non-local instance-id.

Release-Notes: prevent triggering webhooks for events generated remotely
Change-Id: I3a1551e260efa98e716e398465c839c4d43c879e