Support generic HTTP headers in remote requests
Some remote servers require the presence of HTTP headers, e.g. the Authorization header
to supply a token. This change introduces a remote config option "header", allowing
HTTP headers to be added to each request for a remote. Multiple headers can be added in
a similar fashion as the "event" config option.

Change-Id: Ie0f88d8a0776e7637aad863930905c937f55e043