Allow restricting event types using the global configuration
Allow admins to specify which event types can be posted to the
configured listeners. Unrestrictive remote configurations, where
events are not filtered, can cause too many events to be scheduled
on a busy server. With this new global configuration, admins will
have better control over which events can be configured in remote
configurations.

Change-Id: If863ec04eb1fbb0912e638ef9906901957068afb