Process events before scheduling
It's confusing for admins to see tasks scheduled for event types that
listeners haven't configured (or that haven't been globally allowed).
There's very little additional work done in the processing step and the
work saved by not scheduling unwanted events probably more than makes up
for it.

Release-Notes: Skip scheduling tasks for events that should not be processed
Change-Id: I8471a29662d1b88a15e29c5cb6e7e40e25b242f3